feat: create an amarex component for ThemeMaps
