+++ addons/themeMaps/index.js
@@ -1,7 +1,8 @@
-import component from "./components/ThemeMaps.vue";
+import component from "./components/ThemeMapsAmarex.vue";
 import store from "./store/indexThemeMaps";
 
 export default {
   component: component,
   store: store,
 };
+
